refactor(tests): create tests package
